Implemented authentication. Added 'user' field to Todo schema. Minor fixes.

+++ src/controller/todo.js
@@ -1,8 +1,9 @@
-const Todo = require(`../model/Todo`)
+const mongoose = require('mongoose')
+const Todo = mongoose.model('Todo')
 
 exports.findAll = async (_req, res) => {
   try {
-    let todos = await Todo.find()
+    let todos = await Todo.find({ user: res.locals.userId})
 
     res.json(todos)
   } catch (e) {
@@ -12,7 +13,7 @@ exports.findAll = async (_req, res) => {
 
 exports.findById = async (req, res) => {
   try {
-    let todo = await Todo.findOne({ _id: req.params.id })
+    let todo = await Todo.findOne({ _id: req.params.id, user: res.locals.userId })
 
     res.json(todo)
   } catch (e) {
@@ -24,7 +25,8 @@ exports.findById = async (req, res) => {
 exports.create = async (req, res) => {
   const todo = {
     title: req.body.title,
-    body: req.body.body
+    body: req.body.body,
+    user: res.locals.userId
   }
 
   try {
@@ -38,7 +40,7 @@ exports.create = async (req, res) => {
 
 exports.delete = async (req, res) => {
   try {
-    await Todo.deleteOne({ _id: req.params.id })
+    await Todo.deleteOne({ _id: req.params.id, user: res.locals.userId })
   } catch (err) {
     res.status(500).send(err)
   }
@@ -52,7 +54,7 @@ exports.update = async (req, res) => {
       body: req.body.body
     }
 
-    let updated = await Todo.updateOne({ _id: todo._id }, { title: todo.title, body: todo.body })
+    let updated = await Todo.updateOne({ _id: todo._id, user: res.locals.userId }, { title: todo.title, body: todo.body })
     
     res.json(updated)
   } catch (err) {
